+++ src/main/java/svenhjol/charmony/scaffold/mixins/events/client_respawn_player/login_client/ClientPacketListenerMixin.java
@@ -1,0 +1,20 @@
+package svenhjol.charmony.scaffold.mixins.events.client_respawn_player.login_client;
+
+import net.minecraft.client.multiplayer.ClientPacketListener;
+import net.minecraft.network.protocol.game.ClientboundRespawnPacket;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import svenhjol.charmony.scaffold.events.ClientRespawnPlayerCallback;
+
+@Mixin(ClientPacketListener.class)
+public class ClientPacketListenerMixin {
+    @Inject(
+        method = "handleRespawn",
+        at = @At("HEAD")
+    )
+    private void hookHandleRespawn(ClientboundRespawnPacket clientboundRespawnPacket, CallbackInfo ci) {
+        ClientRespawnPlayerCallback.EVENT.invoker().interact(clientboundRespawnPacket);
+    }
+}